+++ templates/javascript/jsdoc/constants.jsdoc.js
@@ -1,3 +1,9 @@
 exports.SCOPE_TYPEDEF = 'typedef';
 
 exports.SCOPE_FUNC_PARAMS = 'func_params';
+
+exports.SCOPE_OBJ_IN_ARRAY = 'obj_in_array';
+
+exports.PARAM_TYPE_PATH = 'path';
+
+exports.PARAM_TYPE_QUERY = 'query';